fix bvh building endless computation bug
Due to floating point inaccuracies it might happen that two triangles end up in the same child node with every split which prevents termination of the algorithm.
Fix it by aborting if no axis is able to split the triangles of a node.
If they are so small and close together there is not really a penalty in doing so.